python_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
deepface==0.0.79
